requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-blob
pandas
numpy
Pillow
opencv-python
albumentations
azure-functions
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu 
torchvision==0.9.0+cpu 